+++ example/judge.py
@@ -7,9 +7,13 @@ from aivle_gym.judge_env import JudgeEnv
 
 class CartPoleJudgeEnv(JudgeEnv):
     def __init__(self):
-        self.env = gym.make('CartPole-v0')
-        super().__init__(CartPoleEnvSerializer(), self.env.action_space, self.env.observation_space,
-                         self.env.reward_range)
+        self.env = gym.make("CartPole-v0")
+        super().__init__(
+            CartPoleEnvSerializer(),
+            self.env.action_space,
+            self.env.observation_space,
+            self.env.reward_range,
+        )
 
     def step(self, action):
         return self.env.step(action)
@@ -17,7 +21,7 @@ class CartPoleJudgeEnv(JudgeEnv):
     def reset(self):
         return self.env.reset()
 
-    def render(self, mode='human'):
+    def render(self, mode="human"):
         return self.env.render(mode=mode)
 
     def close(self):